requirements.txt
keystoneauth1>=3.1.0
python-keystoneclient>=3.8.0
python-glanceclient>=2.8.0
python-neutronclient>=6.3.0
python-novaclient>=9.0.0
python-heatclient>=1.6.1
